requirements.txt
flask>=1.1.2
gitpython>=3.1.18